Edit
Edit ONLY WORKS IF you select a defect,  need to fix that.

As well as, the last portion of quality rep has not been implemented as I'm waiting for instructions on how to go about creating the new NCR and seed data.

Furthermore I tried not to mess up a lot of the designs, and I left some comments near the places I messed up a bit so they're easy to find.