fixed old and new ncr ids
ncr numbers should be fully fixed now.
Just make sure that a single NCR in seed data does NOT have both a new ncr id as well as a old ncr id. It is not implemented in the views to be handled, will adress in the future if thats something we could need.